Remove X509at_get0_data_by_OBJ.
This function's behavior differs from all the other lastpos functions.
It does not appear to be used anywhere, so remove it. (lastpos = -1
returns the first match, lastpos = -2 additionally fails if there are
duplicates, lastpos = -3 additionally fails if the attribute is
multiply-valued.)

Update-Note: X509at_get0_data_by_OBJ is removed. We found no callers of
this function.

Change-Id: I8547bac6626623e43827e2490f04850eb148e317