Remove -2 return value from X509*_get_*_by_NID.
X509*_get_*_by_NID return -1 if the extension was not found, but -2 if
the NID was invalid. Looking through callers, many check index != -1,
rather than index < 0. That means, in theory, they'll do the wrong thing
in some cases.

Realistically, this case is impossible: most callers pass in a constant.
Even in those that don't, NIDs are an local enum, not standard
constants. That means hitting this path is almost certainly a programmer
error. No need to complicate the calling convention for it.

Update-Note: The return value convention of some functions was
simplified. This is not expected to affect any callers.

Change-Id: If2f5a45c37caccdbfcc3296ff2db6db1183e3a95