Sync EVP_get_cipherbynid and EVP_get_cipherbyname
This change syncs the results of EVP_get_cipherbynid and
EVP_get_cipherbyname. Node.js logic assumes that calling
EVP_get_cipherbynid on a NID returned from a call to
getCipherInfo with the cipher name will return the same
cipher information - this assumption holds in OpenSSL
and should also hold in BoringSSL.

This change will allow Electron to remove a patch it is
floating to rectify this mismatch for correct behavior in Node.js.

Change-Id: I40278547b5979e6ff5f370387021c663db3b4720