WIP: Generates "low-level" bindings for Rust using bindgen
This is designed to be the minimal infrastructure required to support
using BoringSSL in the Rust ecosystem without fear of ABI drift. Bindgen
is used to generate Rust bindings in lockstep with the rest of the
build. `rust-openssl` can consume these generated bindings with minimal
changes.

There are a few "hand-rolled" constants and aliases that are not exposed
through bindgen that are added for upstream compatability.

Change-Id: I1dacd36a4131e22a930ebb01da00407e8465ad7e