Add magic tag to BoringSSL binaries.
There are cases where people grep binaries for strings like OpenSSL
version strings in order to detect when out-dated versions of libraries
are being used. With BoringSSL you might find "OpenSSL 1.1.1
(compatible; BoringSSL)", if the linker didn't discard it, but that's
not very helpful for knowing how up-to-date BoringSSL is because we
hardly ever change it.

This change adds a distinct random value to search for that uniquely
identifies BoringSSL and includes a rough guide to how old the BoringSSL
copy is. The linker will hopefully not discard it because it's in the
libcrypto init function, and a volatile pointer is used to try and get
the compiler not to discard it.

Change-Id: Ie2259fd17a55d249a538a8a161b0d755396dd7b8