Specify ANDROID_ARM_MODE in 32-bit ARM builders.
This resolves the TODO added in
https://boringssl-review.googlesource.com/c/boringssl/+/45544, though
it's only a temporary solution. An even future NDK will make this
impossible altogether. (See bug.)

I've specified this via the cmake_flags property, though most of the
other CMake options are still hard-coded into the recipe. I still need
to go shift all the CMake parameters into this file, so we can specify
bots without having to update the recipe so often.

Bug: 454
Change-Id: I31d81583a690acaec3ccc4df4f6510771d95aabd