Enable Python 3 recipes differently.
The original instructions for enabling Python 3 don't work because they
depend on some other thing being enabled, and some things are still on
the old thing. Use this other way to use Python 3 recipes, which is
supposed to avoid this problem with the other thing.

(I'm told the relevant things are named "kitchen" and "bbagent".)

Bug: chromium:1256384, chromium:1256383
Change-Id: I0b2d12003d2525f461a5830266d9bcf98a9b6752