Sync sha512-armv8.pl up to 753316232243ccbf86b96c1c51ffcb41651d9ad5.
This imports 753316232243ccbf86b96c1c51ffcb41651d9ad5,
46f4e1bec51dc96fa275c168752aa34359d9ee51, and
32bbb62ea634239e7cb91d6450ba23517082bab6.

The last commit fixes a detection of big-endian aarch64 in the kernel,
which we do not support at all, but is imported to reduce the upstream
diff. Though it points out a messy part of arm_arch.h: __ARMEL__ and
__ARMEB__ are specific to 32-bit ARM. __AARCH64EB__ and __AARCH64EL__
are the 64-bit ones. But OpenSSL's arm_arch.h defines them for aarch64
too and most of perlasm relies on this. We should fix the files upstream
to do this consistently.

Change-Id: I6c2241e103a97e8c3599cdfa43dcc6f30d4a2581